refactor(api): extrai a black-friday de products para modulo dedicado
Move a logica da campanha (servico, controller e rota) para arquivos
blackFriday.* proprios, tirando-a de products.*. O app principal segue
montando /api/v1/products/black-friday, entao o comportamento externo
e identico; e so um passo de organizacao antes de isolar a campanha.

As faixas de preco validas agora derivam de PRICE_RANGES no servico,
eliminando a lista duplicada que vivia no controller.
