+++ backend/src/controllers/products.controller.js
@@ -29,28 +29,6 @@ function getFeaturedProducts(req, res) {
     res.status(200).json({ products, total: products.length });
 }
 
-const VALID_PRICE_RANGES = ['all', '0-50', '50-100', '100-300', '300-999'];
-
-function getBlackFridayProducts(req, res) {
-    const { category, priceRange, minRating } = req.query;
-
-    if (priceRange !== undefined && !VALID_PRICE_RANGES.includes(priceRange)) {
-        return res.status(400).json({ error: 'priceRange invalido' });
-    }
-
-    const result = productsService.getBlackFridayProducts({
-        category,
-        priceRange,
-        minRating,
-    });
-
-    res.status(200).json({
-        products: result.products,
-        filters: result.filters,
-        total: result.products.length,
-    });
-}
-
 function getRecommendations(req, res) {
     const raw = typeof req.query.cartItemIds === 'string' ? req.query.cartItemIds : '';
     const cartItemIds = raw
@@ -104,7 +82,6 @@ module.exports = {
     getProductById,
     searchProducts,
     getFeaturedProducts,
-    getBlackFridayProducts,
     getRecommendations,
     createProduct,
     updateProduct,
