+++ backend/src/routes/products.routes.js
@@ -7,7 +7,6 @@ const router = express.Router();
 router.get('/', productsController.listProducts);
 router.get('/search', productsController.searchProducts);
 router.get('/featured', productsController.getFeaturedProducts);
-router.get('/black-friday', productsController.getBlackFridayProducts);
 router.get('/recommendations', productsController.getRecommendations);
 router.get('/:id', productsController.getProductById);
